requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
DateTime==4.8
discord==2.1.0
discord.py==2.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
zope.interface==5.5.2
